finish local format and short cut, that is awesome
but!!! mongodb relation is not finish yet!!!!

+++ local_format.py
@@ -12,11 +12,14 @@ video_ext = ['mp4', 'avi', 'mkv', 'wmv', 'mov', 'mpeg', 'rmvb', 'mp3', 'flv', 'q
 img_ext = ['jpg', 'jpeg']
 
 format_root = os.path.join(format_path, "all")
+undefined_root = os.path.join(format_root, '_Undefined')
 format_role = os.path.join(format_path, "role")
 format_score = os.path.join(format_path, "score")
-for v in [format_root, format_role, format_score]:
+format_images = os.path.join(format_path, "shortcut_img")
+for v in [format_root, format_role, format_score, undefined_root]:
     if not os.path.exists(v):
         os.makedirs(v)
+
 
 def findVideo(name):
     video = None
@@ -61,9 +64,10 @@ def createFormatPath(video, root=format_root, format=["code", "_id"]):
 def moveAFile(video, file_path):
     target_path, relpath = createFormatPath(video)
     shutil.move(file_path, target_path)
-    shutil.copy(os.path.join(img_path, video.img), target_path)
+    if os.path.exists(os.path.join(img_path, video.img)):
+        shutil.copy(os.path.join(img_path, video.img), target_path)
 
-    Video.objects(_id=video._id).update_one(set__local_path=target_path)
+    # Video.objects(_id=video._id).update_one(set__local_path=target_path)
     return target_path
 
 
@@ -76,7 +80,7 @@ def moveAImage(video, file_path):
 
 
 def shortcutBuild(video):
-    target_path = createFormatPath(video)
+    target_path, relpath = createFormatPath(video)
 
     # role base
     for role in video.roles:
@@ -84,7 +88,7 @@ def shortcutBuild(video):
         if not os.path.exists(path):
             os.makedirs(path)
         winshell.CreateShortcut(
-            Path=path + video['_id'] + ".lnk",
+            Path=os.path.join(path, video['_id']) + ".lnk",
             Target=target_path)
 
     # score base
@@ -93,8 +97,13 @@ def shortcutBuild(video):
     if not os.path.exists(path):
         os.makedirs(path)
     winshell.CreateShortcut(
-        Path=path + video['_id'] + ".lnk",
+        Path=os.path.join(path, video['_id']) + ".lnk",
         Target=target_path)
+
+    # images
+    target_path = os.path.join(img_path, video.img)
+    if os.path.exists(target_path) and not os.path.exists(os.path.join(format_images, video['_id'] + ".jpg")):
+        shutil.copy(target_path, os.path.join(format_images, video['_id'] + ".jpg"))
 
 
 def loadAPath(dir, video_type=video_ext, img_type=img_ext):
@@ -106,18 +115,18 @@ def loadAPath(dir, video_type=video_ext, img_type=img_ext):
             ext = f[f.rfind('.') + 1:]
             if ext in video_ext:
                 # find target to move
-                name = os.path.splitext(f)[0]
-                video = findVideo(name)
+                video = findVideo(f)
                 if video is not None:
                     video_count += 1
-                    moveAFile(video, os.path.join(root, f))
+                    moveAFile(video, os.path.join(root.decode('gbk'), f.decode('gbk')))
+                else:
+                    shutil.move(os.path.join(root.decode('gbk'), f.decode('gbk')), undefined_root)
                 pending_files.remove(f)
             elif ext in img_ext:
                 # find key code img to move
-                name = os.path.splitext(f)[0]
-                video = findVideo(name)
+                video = findVideo(f)
                 if video is not None:
-                    moveAImage(video, os.path.join(root, f))
+                    moveAImage(video, os.path.join(root.decode('gbk'), f.decode('gbk')))
                     pending_files.remove(f)
 
         files = pending_files[:]
@@ -125,18 +134,46 @@ def loadAPath(dir, video_type=video_ext, img_type=img_ext):
         if video is not None and video_count == 1:
             for f in files:
                 if f[f.rfind('.') + 1:] in img_ext:
-                    moveAImage(video, os.path.join(root, f))
+                    moveAImage(video, os.path.join(root.decode('gbk'), f.decode('gbk')))
                     pending_files.remove(f)
 
         if len(pending_files) == 0 and len(dirs) == 0:
-            shutil.rmtree(root, ignore_errors=True)
-            print "delete?"
+            # shutil.rmtree(root, ignore_errors=True)
+            print "delete: " + root.decode('gbk')
 
-# format videos
-for path in wait_format_path:
-    loadAPath(path)
 
-# build shortcuts
-videos = Video.objects(local_path__exists=True)
-for video in videos:
-    shortcutBuild(video)
+def formatVideos():
+    # format videos
+    for path in wait_format_path:
+        loadAPath(path)
+
+
+def updateLocal():
+    # scan local files and set into video
+    # Video.objects(_id=video._id).update_one(set__local_path=target_path)
+    # format_root > files > ids
+    Video.objects().update(unset__local_path=True)
+    all_codes = os.listdir(format_root)
+    for codes in all_codes:
+        codePath = os.path.join(format_root, codes)
+        all_code = os.listdir(codePath)
+        for code in all_code:
+            target_path = os.path.join(codePath, code)
+            Video.objects(_id=code).update_one(set__local_path=target_path)
+
+
+def buildShortcuts():
+    # build shortcuts
+    shortcut_dirs = [format_role, format_score, format_images]
+    for d in shortcut_dirs:
+        shutil.rmtree(d, ignore_errors=True)
+        os.makedirs(d)
+
+    videos = Video.objects(local_path__exists=True)
+    for video in videos:
+        shortcutBuild(video)
+
+
+formatVideos()
+updateLocal()
+buildShortcuts()
